Refactor ListParticipantsHandler a bit
Preload usernames and User objects immediately, so that we can have a
single foreach loop for all participants. The array of participants can
have 20 elements at most, so the array_map is still very cheap and we
don't need the double loop optimization.

Break the code down into a couple functions.

Change-Id: I605460d8dc8b51df752f4c50a0b7e08583c0b105