Add maintenance script for invitation list generator prototype
This patch introduces a prototype of an invitation list generator,
implemented as a maintenance script. The script takes a list of articles
as input; it then grabs all the contributions to those articles for the
last 3 years, and assigns a score to each contributor based on the size
of their contributions, their overall edit count, and their recent
activity.

The input pages can be from multiple wikis, but this uses an ad-hoc
syntax (as opposed to interwikis) due to limitation in the title parser.
For the same reason, all pages must be in the mainspace.

Note that this is meant as just a proof of concept. The pseudo-models
used here are not based on real-world data, and are just a very rough
approximation of how a human organizer might analyze user contributions
when looking for contributors to invite.

In general, this code is very experimental and somewhat hacky, due to
its prototypal nature. It may be changed, or removed altogether,
depending on user feedback.

Bug: T353459
Change-Id: Ibbf7e3a79e4e658d3de822c0db4633f260f6b797