Add footer to organizer emails
This is similar to the footer used for emails sent through
Special:EmailUser, but provides more specific information about the
event.

Special:Mute is not the ideal way to manage these emails, but we're
using it as an interim replacement while we think about potentially
creating a dedicated email management system.

Bug: T348043
Change-Id: Iee221be129472321e6ce6f99411253800dd4e25c