Do not show clickwrap form if there are no answers
If there are no answers, or if answers still need to be aggregated,
accepting the clickwrap is pointless because the user still won't be
able to see anything (ever if no answer, temporarily if pending
aggregation).

So, check the presence of answers before checking for clickwrap
acceptance, and only show the clickwrap form if answers are ready to be
displayed.

Bug: T355128
Change-Id: I737dcfa028ac49c80e22cf1c9a30689d82aab402