Suppress phan errors in REST handlers
Starting with core change I1b002b4dbd, phan knows that the return value
of getValidatedBody() is nullable, and complains when it's not treated
as such. However, it looks like it should never be null under normal
circumstances, as written in T357909. So, just suppress the phan errors
until this can be fixed properly in core.

Change-Id: I3870119dfbe465c88ee0d7e311eab6612baea238