Allow organizers to choose questions when enabling registration
Organizers can now choose which questions to enable for an event, when
they are enabling registration.

Currently, the selection cannot be changed when editing an existing
event. This will be changed in a subsequent patch.

Also update the form to match the latest prototype. This includes
updating a few messages, as well as removing the section for the
clickwrap agreement, that is now only shown in the statistics tab of
Special:EventDetails.

Bug: T322209
Change-Id: I4f323bb211ab8da240bfc7c047ce919a56ce38dc