Use canonical links in organizer email footer
The full URL does not include the protocol (it uses PROTO_RELATIVE by
default), and some email clients fail to recognize it as a link.

So, introduce a new getCanonicalUrl method, akin to
Title::getCanonicalURL, which returns a canonical version of the URL
that includes the protocol and is generally guaranteed to be suitable
for email notifications and the like.

Bug: T348043
Change-Id: I9738ebc1617f41b921426b34faec5b61ba822f4a