GenerateInvitationList: Improve debug logging
Include the overall batch index, as well as the per-page one, report the
total number of overall batches, and do not include the start values for
the first batch.

This is far from perfect, and I believe we should look into more
accurate progress markers (e.g., by counting the revisions) when making
this code accessible to users.

Bug: T357007
Change-Id: I8d6394456ed5741277b36e39a56448e695348284