Disallow moving event pages outside of the Event namespace
This isn't really supported, and therefore it won't properly.

It looks like this should be a super-rare occurrence in practice, as
demonstrated by the absence of pages outside the event namespace in both
beta and production. Therefore, just prevent the move.

This will be reconsidered if we decide to allow event registration to be
used on pages outside the Event namespace.

Add a test for this, to make sure that moving pages outside of the Event
namespace is not allowed if they have an associated registration. Note
that the test needs to be made an integration test because it uses the
NS_EVENT constant.

Bug: T358704
Change-Id: I39a8f0a3969e97cf62a771fa7ba770d248c08dad