GetOwnRegistrationInfoHandler: mark answers to removed questions
In the GetOwnRegistrationInfoHandler response, add a new property to
indicate that a certain answers is for a question that has since been
removed from the event. The client can then decide what to do about it.

This information is not currently being displayed anywhere, but it will
be as part of T358529.

Bug: T358529
Change-Id: I7a39d3658996b4459c22a1efbef9403b190b10ab