Always show answered questions when a user registers for an event
When someone registers for an event (via the special page or the form on
the event page), we would previously only show the currently enabled
participant questions for that event. With this patch, we show both the
enabled questions AND any questions that the user already answered in
the past; these are questions that have since been removed by the
organizer.

This allows participants to edit their information even if the organizer
removes a question.

If the user blanks out their answer to a question that has since been
deleted, the field for that question will no longer be shown in either
form. Note that this is currently affected by a minor bug in
Special:RegisterForEvent, where said fields are still visible in the
form after the initial submission (but they go away when refreshing the
page). Fixing this seems unnecessarily complex for the time being.

Bug: T354880
Change-Id: Idbe013caaf4364edac8a997fb9c824ef79216ee5