Rework MWEventLookupFromPage to be the main event-by-page lookup
Previously it was just a convenience helper for MW classes. With this
patch, it becomes the main class to use when performing a lookup by
page. One of its two entry points still accepts MW-specific classes,
whereas the other is for the ICampaignsPage interface.

The class doesn't do anything special at the moment, but the idea is
that it can be used to canonicalize the given page. For instance, a
subsequent patch will make it so that when a translation page is given,
the actual lookup will be done on the source page.

Rename and re-namespace the class to highlight the fact that it's no
longer just another bridge between MW and campaigns code.

Use the new lookup in EventPageDecorator and EditEventCommand, and
update tests accordingly.

Bug: T357716
Change-Id: Idff378dce6e64419b50484a5706fde0e792f54e4