Make translated event pages use the same registration as the source page
Any subpage of a translatable page, being just a translation, should
still use the event registration of the source (parent) page.

This is achieved by canonicalizing the given page in PageEventLookup
before hitting the DB.

Note that this can only be done if the given page is local, because
Translate doesn't support checking foreign pages. This should not be a
problem in practice.

Also adjust some typehints to use PageIdentity instead of
ProperPageIdentity, as the latter isn't widely used in typehints, but
they're pretty much the same thing.

Bug: T357716
Depends-On: Ib03ed861d95f5c042ed63291ef539d84b7db7077
Change-Id: I6d90fa3ee7969b48ecd1fea2afc7c9123c621e74