Streamline ICampaignsPage instantiation
Introduce a new method in CampaignsPageFactory to convert from MW page
interfaces (PageIdentity and LinkTarget), and use it in places where we
were constructing MWPageProxy directly. This avoids direct dependencies
on other MW interfaces (such as PageStore and TitleFormatter) and
encapsulates knowledge of the $prefixedText parameter (which is just
an implementation detail).

Change-Id: Id9646cae74d12345c1b82c7139c48a6e693ce938