AggregateParticipantAnswers: prune old aggregates
As part of T354880, we want to delete obsolete aggregates, when one or
more questions are removed from an event; this will only happen once the
event has ended, so that the organizers can choose to re-enable the
questions before the event ends, if they wish to.

So, add a new method which runs after the main aggregation and goes
through all the past events to find those with aggregates to delete. As
noted in the code, this cannot be done in the same code path that
aggregates new answers.

Bug: T354880
Change-Id: I6b485099d259ec0a6376934fca18337afa5d2f5a