PageMoveAndDeleteHandler: skip canonicalization of event pages
When checking whether a page is being moved or deleted, we must not
perform any canonicalization. Otherwise, whenever a non-canonical page
is moved/deleted (e.g., a translation subpage), the event would be
renamed or deleted as well.

So, add a parameter to PageEventLookup that determines whether the page
should be canonicalized.

Bug: T362365
Change-Id: If23cf3397cc03ea694d08008dcf4852b886718ac