EventKebabMenu: fix promise chaining for the 'delete' action
jQuery's `.then()` always returns a new promise to allow chaining. It
never returns the original promise, not even if a handler has no
`return` statement. When the deletion fails, we want the rejected
promise to bubble up to the caller (`onChooseOption`), so that it can
fall back to a special page; however, the rejected promise was being
discarded by `then`, and the caller would always see a resolved promise.

So, use `.fail()` to log the error, so that the original rejected
promise remains untouched.

Change-Id: Ic922ca093365959a00196df2bc641a17e7b5515e