Split EventsListPager styles out of EventsPager.less and rename
The original idea was to have the styles for both pagers in the same
place, even though they apply to different things. This proved to be
challenging given CSS's nature of being cascading, as well as the
no-descending-specificity rule enabled in our stylelint configuration.
Because there rules require the styles to be ordered by specificity, it
becomes impossible to group them by page that they apply to.

So, have a stylesheet for each pager, and rename them according to the
PHP class name of the pager that they are for.

Follow-up to I331bf77b24a.

Change-Id: I92a7c57993a1557a8f2973314dd8925ba4fcbbea