Check deleted participants records for aggregated answers
When someone intends to register for an event (via the event page
dialog, the API, or Special:RegisterForEvent) we check whether their
answers have already been aggregated, and don't let them answer again if
that is the case.

However, this check was previously performed on the "current"
participant data, which only exists if the user is currently an active
participant.

But if someone registers for an event, answers questions, then those
questions get deleted after 90 days, and they cancel their registration,
then they would be able to answer the questions again when registering
for the second time.

So, add a new method to ParticipantsStore to check whether a given user
has ever had any answers aggregated. This check is also performed on
non-live records, unlike `getParticipant()`.

Bug: T363193
Change-Id: I0d1f601338c5b7d17d4f8074474690bad82c9f81