Re-enable previously failing selenium test
Apparently, there's a brief moment when the menu entry exists in the DOM
but is not clickable. Wdio tries to click it but fails, hence resulting
in a test failure (not deterministic, and it also depends on the
device speed; hard to reproduce locally, but fails a lot in CI).

So, make sure that the menu entry is visible before trying to click it.
This is based on AdvancedSearch's
tests/selenium/pageobjects/search.page.js which also uses a
TagMultiselectWidget (https://w.wiki/9vAJ).

Also add specificity to the selector, and note that there are no
extension-specific classes in the whole menu hierarchy.

The test is now passing 100/100 times in CI (see Ic4314d34ccd9).

Follow-up to I5c9d73a2a84.

Change-Id: I13f86c17881b84405f248ae8fcca773e195188f0