Fix spacing issue on Special:AllEvents
For events with a long name, the event info container grows as much as
it can to try and fit the title. In turn, this causes the day container
to shrink and breaks vertical alignment.

So, use `min-width` instead of `width` on the day element to prevent it
from shrinking (functionally equivalent to `flex-shrink: 0`).

Bug: T353382
Change-Id: I78bae88603076e2dd2e5d2775c3a4d8e569a845c