SpecialEventDetails: fix chat and meeting links on foreign wikis
Have EventDetailsModule accept a single parameter for the wiki ID, and
then compute any derived data if and when it actually needs it.

The previous signature seemed to imply that the foreign details URL was
always set, but in fact it was only set if the viewing user is an
organizer.

Bug: T324252
Change-Id: I33f2aa7e2fa19193c1153feabb33948a70be2f2f