Add batching to various queries in EventsListPager for performance
Pull all the creators, extra organizers, and organizer counts at once
when preprocessing the results, instead of per-event. Also run a single
query to get all the usernames (even if we may end up not using all of
them), and use LinkBatch to preload all the user page links.

On my local setup, where most events have the same organizer (and
therefore the impact is not too tangible), viewing a list of 500 events
now results in a total of 127 queries, down from 1642.

Bug: T366828
Change-Id: I84adb08eb547c726e3913da953f79340d7740dfc