Move the invitation list generator code to a standalone service
This patch moves the existing invitation list logic to a standalone
service, which is in turn used by the maintenance script. The script
is kept, so that it's still possible to generate one-off invitation
lists via CLI.

The new service will also be used by other backend classes to generate
invitation lists at the user's request. This will be implemented in
subsequent patches.

Note that this patch only makes minimal adjustments to the generator
code. Other things such as a dedicated Job class will also be introduced
in subsequent patches.

Bug: T365628
Change-Id: I0438adbeab2542a523aff3d85975e5c575f9f3bf