Introduce Worklist entity and parser
Introduce a new Worklist entity to wrap the array of PageIdentity
objects that we've been using. This allows more type safety and avoids
the repetition of the detailed phan types. It will also easily let us
define serialization rules for this class, in the context of job
parameters.

Factor out the worklist parsing logic from PotentialInviteesFinder, so
that it can be independently used to validate a worklist input. Note
that the validation errors are still raw exception, and not localised,
user-facing errors. This will be changed in a future patch, as part of
T365629.

Also clear some confusion in the handling of wiki IDs (string vs false),
with heartfelt thanks to PHP for making it more complex than it should
be.

Bug: T365628
Bug: T365629
Change-Id: I6e4c5c5655cccaf9d027e20cfc6760f0828e290f