Introduce InvitationListGenerator
This class sits in the behaviour layer, between the interface (special
page) and the storage layer (implemented in I9a4a4d9104). It does
permission checks, as well as validation of the event page.

For now it is just a stub, since there's no storage and nothing is using
it for now.

Bug: T365629
Depends-On: Ibe3dd044634bf5a68504131a543147cdc496bbb0
Change-Id: I9f8ba81e0fc480c7b9f8f774f799ce443f4126ce