Rename InvitationListGenerator and related classes
In our current plans, "invitation list" refers to an entity which
comprises general information (name, wiki, creator, ...), a worklist,
and a list of potential invitees. This service, however, is only
responsible for this last part, and the new name more accurately
reflects that.

The previous name would also get in the way of creating new classes that
deal with invitation lists as a whole, such as a class for validating
and creating an invitation list.

As always, naming is hard, and alternative naming proposals are welcome.

Bug: T365628
Change-Id: I1c1f9916eaa69a698b95cc3b99ea9c9531e446ca