Generate and store invitation list when special page form is submitted
Make InvitationListGenerator use the new storage abstraction to store
the invitation list and worklist when the form is submitted
successfully. Also enqueue the job that will generate the list of
potential invitees.

Use the correct name for the job class in extension.json.

Redirect the user to Special:InvitationList upon successful submission.

Bug: T366641
Change-Id: If26e947c98253716efca3ba8dc27832767b14428