Rename Community List-related code
Hooks should not know about their implementations, so use a generic name
for it, that describes when it's being fired.

Rename CSS selectors that are in this repo for performance reasons.

Add a temporary class alias to please CI during the rename.

Please merge together with Id766c8e00cb2d821d8f6514713447560ef829fbd.

Bug: T376218
Depends-On: I4cc346cda9e17864ee24c200b0dae83d9ea9cbba
Change-Id: I875dae0d69bedd850437c7a4f87741d2b85ac257