Redirect user to event page after enabling registration
Instead of showing the success/warning messages on the special page,
redirect the user to the event page. With JavaScript enabled, those same
messages will then be rendered as bubble notifications.

Use session data to store whether registration has been enabled and any
warnings that came with it; specifically, do not use URL parameters that
would re-trigger the notification upon every subsequent visit, including
from people other than the organizer.

The behaviour when updating registration is unchanged for now, but the
code should be easily extensible to that use case as well. In that
regard, do note the usage of "update(d)" and not "enable(d)".

Bug: T362924
Change-Id: I346c5740b8b5c44cef823029d1448d1e384b26a2