Update text of registration enabled notification
The main difference is that this notification is now shown directly on
the event page (as opposed to a special page), so the message no longer
includes a link to the event page.

Since we're changing the format, create a new message instead of
updating the old one. This also allows us to conform to the naming
scheme used for other messages on event pages.

Bug: T377998
Change-Id: Ia12978181df71a1fbf5aff18550111bbfc70c01a