Add internal REST endpoint to format date strings
This allows us to format a date-time according to the user preferences,
in all the formats supported by Language::sprintfDate(). As discussed in
T21992, this functionality is not exposed in JS (as the PHP
date-formatting logic is quite complex and supports loads of languages).

So, introduce an internal REST endpoint to provide indirect access to
sprintfDate in the frontend.

Bug: T362639
Change-Id: I2b1f0363a2c6cc7f22dc01bd6cfd0ef18cc32ae5