Add TimeZoneConverter JS utility
This can be used to format a date range with timezone indication
according to the user's preferred timezone as specified in the browser
preferences.

This patch only introduces the new utility, and loads it in both the
event page and special pages. However, it is not being used anywhere
yet. Note that module base directories need to be updated to accommodate
for this file that lives outside of the module directory.

Add QUnit boilerplate, as well as a QUnit test for this new class.

Update the specialPages module dispatcher so that it doesn't emit a
warning when loaded in QUnit (there's no page specified there).

Note that, while we use modern(-ish) Intl classes to get the user's
preferred timezone, we still need to use the less modern `moment`
library to get a formatted timestamp that can later be formatted,
according to the user MW preferences, on the server-side. This is
motivated by the fact that it's apparently impossible, using just plain
JavaScript, to "apply" a timezone to a datestamp, and then format it
arbitrarily. All we can do is format it according to a given locale, but
then it wouldn't be possible to re-parse it on the server side for the
final formatting.

So, use momentJS to get an intermediate representation, which is then
used by the server to give us a formatted localised string.

Bug: T362639
Change-Id: If2ff7cd0dabc8b4267ed35579b73ed0f79ee9642