Add event wikis to EventRegistration and EventStore
Add a new property to EventRegistration to the event wikis, set in the
constructor, and a getter. This can be either a list of wiki IDs, or a
special value to represent "all wikis". Introduce a new constant for
that, to be used outside of the storage layer. Using `true` instead of
a string allows better static analysis, because in code such as:

```
if ( $wikis === EventRegistration::ALL_WIKIS ) {
    // Do something
} else {
    // Do something else with the list of wikis
}
```

it's easy to infer that $wikis can only be an array in the `else`
branch. If we used a string, the value could still be a string in the
`else` branch (phan supports literal strings in union types, but I don't
think that's standard, and `true` as a standalone type is more likely to
be recognized by tools). Rename the stored constant to clarify that it
should only be used by the DB layer.

Update EventWikisStore accordingly, and introduce a multi-lookup method
to retrieve the wikis for multiple events; the single-event lookup is
now a special case of the new method.

Add a dependency on EventWikisStore to EventStore, and use it to
read/write wikis in various places.

Update tests and everthing else that is affected by the signature change
to the EventRegistration constructor. In particular, update a test in
EventWikisStore that wasn't using its data provider.

Note that EventFactory is not updated in this patch, and more generally
there is no way for users to access or update the stored wikis. This
will be implemented in another patch.

Bug: T379276
Change-Id: Ia4010f8b32792aac2f924d2d9dbd0cc714d959de