Add support for event wikis to the API endpoints
Introduce a feature flag to control whether the event wikis feature is
enabled.

The "enable-" and "update event registration" endpoints now support
setting the wikis of the event when the feature is enabled. The wikis
can be specified as a list of wiki IDs, or using the special value "*"
as the only element of the array, consistent with the format used by
ParamValidator for enum parameters. This is then converted to the
internal array|true format later in the handlers, if the validated list
contains all wikis. Note that this has an edge case where in a small
wikifarm (<100 wikis), if someone selects all wikis manually, we would
convert that to the special "all wikis" value. This is just an edge case
though, and not enough to prevent us from using the builtin PARAM_MULTI
syntax.

In the "get event details" endpoint, return the list of wikis if the
feature flag is enabled. This uses the same format as the write
endpoint, so that values from the GET endpoint can be used immediately
to issue a write request.

When the feature flag is disabled, the behaviour is unchanged: it is not
possible to specify any wikis, and they're not included in the response
of the GET endpoint.

Add `array_unique` to WikiLookup::getAllWikis(). MediaWiki core does not
enforce uniqueness, and duplicated wikis are possible (but harmless) in
case of misconfiguration, see for example
https://www.mediawiki.org/wiki/Special:Diff/6854472.

Bug: T379286
Change-Id: Icfa05086af1114fdb7e78d4edf58cb8c99698855