Implement wiki selector when enabling/editing registration
Add a radio type to let users choose whether the event focuses on
certain wikis, and a multiselect for choosing specific wikis. The field
is displayed as a subfield with ad-hoc CSS.

Flip the array returned by WikiLookup::computeListForSelect, as the
format used by the multiselect widget has the display names as keys, not
values.

Add custom CSS to limit the size of the multiselect when serving the JS
version, so that we don't risk showing a huge list of 1000 checkboxes
while we're waiting for infusion.

Bug: T367226
Change-Id: If74958d78b8c8adea869c1f7688df7c3c7a6ea88