Fix broken PHPUnit test
With core commit I7e50be91c01b4b70e1d1380651f001e35bd60523,
IMessageFormatterFactory::getTextFormatter() has a typehint of `string`
on the language code parameter. In GetParticipantQuestionsHandler, we
retrieve the language code from the content language object, via
`Language::getCode()`. However, the latter method does not have a return
typehint, and it was not explicitly configured in the test. As a result,
PHPUnit would make the mock return `null`, which breaks the new
typehint.

So, explicitly configure the `getCode` method so that it returns a
string.

Change-Id: I4d28a157aaeb4faa17f4785dccd3b591230b03d0