Rest: fix handling of "all wikis" when enabling/editing registration
Comparing the arrays directly is sensitive to the order in which wikis
are specified in the request body. So, compare the counts instead.
Because one of the operands is the list of all wikis, it is guaranteed
to be a superset of the provided list of wikis (already validated), and
therefore we don't need to check that their elements are also the same.

Note that this is mostly only relevant when testing locally, with a low
number of wikis that can be specified individually in the request body.
In a production-like environment with hundreds of wikis, the only way
for the two lists to be equal is using the special value `"*"`; but that
gives us the same ordering on both sides of the equality.

Bug: T379286
Change-Id: If718ef1dabc2b6ca0ea02af9e6034a336bda629f