selenium: Make "cancel registration" test more robust
First, make sure that the registration was successful before trying to
cancel.

Next, if we just registered, wait for the notification to appear and
then make it go away before trying to unregister. See code comment for
detailed rationale, but the TLDR is that the notification overlaps with
the menu and makes it not clickable, which makes puppeteer throw an
unrecoverable error.

Change-Id: If6f47103f788a6dbef757c04c51e42fa7a31606d