RegisterParticipantCommand: upstream edit check
Edits should always be allowed when an event has ended or registration
is closed. This is a general truth. Therefore, it makes sense to have
this check in the general `checkIsRegistrationAllowed` method, instead
of special-casing it just for the command in `registerUnsafe`.

So, add a parameter to `checkIsRegistrationAllowed` to perform this
additional check, and class constants for this parameter. Update callers
accordingly -- the existing usages outside of RegisterParticipantCommand
really only need to check for new registrations.

Note that this introduces a minor behaviour difference in the
registration workflow: errors due to the user does not having a global
account take now precedence over errors due to the event being deleted.
This is highly unlikely to happen in practice though, and so presumably
acceptable.

Add a unit test for `checkIsRegistrationAllowed`.

Bug: T311403
Change-Id: I084b595357938074bc394eeac7a9208447dd76f8