selenium: Make myEvents.js more robust
This test is currently brittle because it simply deletes the first event
in the list, whatever that might be. In practice, this means that:
- In a wiki farm environment, the first event on the list might belong
  to a different wiki, and therefore can't be deleted.
- When running tests concurrently, we might delete an event that is
  being used by another test.

So:
- Add a method to myEvents.page.js to allow filtering events by name
- Add a test to make sure that said method works
- Filter events by name, using the name of the event we just created, so
  that we know what event we are deleting.
- Improve the selector for the manage button, so that it does not rely
  on an implementation detail (the icon), and that it waits for the
  button to be infused and become a ButtonMenuSelectWidget.

Note that the updated code is assuming that the newly-created event will
be the first match when searching by name. In practice, this is highly
likely because of the randomization introduced in `getTestString`.

Change-Id: I54142488c9c6784175ab3d27d7377a06cd99548d