Avoid deadlock-prone `SELECT ... FOR UPDATE` queries
When `SELECT ... FOR UPDATE` is used on a row that does not exist, MySQL
is unable to lock the nonexistent index position. Instead, it will issue
a next-key gap lock up until the first empty position following the
selected record.

However, if the table is almost empty or it contains sparse data, there
could easily be no "first empty position" to lock, in which case the
supremum record will be locked. In turn, this prevents any new
insertions on a semi-empty table, because the whole index is basically
locked.

So, remove all `FOR UPDATE` statements from queries that may not find
any matching records. We are already reading data from the master DB,
thus providing defense against replication lag. In the rare case of
simultaneous updates, we are fine with "last request wins".

Bug: T381622
Change-Id: Id13840873ba95952ac6213d0615148cfb6250bf4