Set explicit meeting type default in the registration form
OOUI radio inputs in HTMLForm do not support an indeterminate state:
when no valid default is given, it just autoselects the first available
option (T207475).

However, relying on this behaviour seems ill-advised, because it
prevents other things from working correctly, like the hide-if detection
introduced in I5ed4951ce97d4c.

So, make the "event meeting type" field explicitly default to "online"
(the first option). Note that we might reconsider this decision when
switching to Codex, if that allows indeterminate radios.

Change-Id: Iba43899eec025a9e00d7fb3d20d1a54c2ce551f1