Add length limits to country and address fields
These are intentionally very permissive, and not meant to be hit under
normal circumstances. Consequently, the implementation favours
simplicity over correctness (e.g., there's a discrepancy in bytes vs
codepoints for the UI).

The main purpose of this change is not to prevent mistakes, but rather
to prevent a scenario where, if an address is too long, the country will
not be stored at all in EventStore::updateStoredAddresses(), where it is
appended to the address after `" \n "`. This would result in invalid
stored data, and uncaught exceptions upon retrieval.

Bug: T325380
Change-Id: If83847ec9a4c2026844852dc5ef70dba6d3cf1e4