SpecialEditEventRegistration: re-order permission checks
Check whether the event is deleted or non-local before checking
permissions. This allows us to show a consistent message when the event
is not local. The message was previously implemented in Iacc371fad6ced2,
but locality was checked after permissions; consequently, it regressed
in I8320fc15125b69f6, when locality checks were added to
PermissionChecker.

So, postpone permission checks and show more specific error messages
when possible.

Note that, as a side effect of this change, for non-local events we
always show the "edit event on XXwiki" message, even to non-organizers.
However, performing the full permission check on a foreign wiki is
relatively hard (and really only possible via foreign API requests),
which is why we're not allowing cross-wiki edits in the first place.

Bug: T366969
Change-Id: I5be6e33ae757847cfdb6aa84045540a5a8ca8faf