selenium: wait for "more details" handler to be installed before clicking
On a slower machine, the button might be clicked before the handler is
installed. Because the button is actually a link (that gracefully falls
back to Special:EventDetails), this results in the special page being
opened and the assertion running there (where it fails).

So, wait for the listener to be installed, by using the same code
already used for the "register" button, that is now moved to its own
helper function.

Bug: T383336
Change-Id: If298bc27d9e0d2b88db5145c04ac8bb96ba1fecf