Improve type safety for date filters in EventsListPager
Make the $startDate and $endDate parameters nullable, and use null
instead of the empty string to indicate the lack of a value.

Validate timestamps upfront and immediately reject invalid ones, rather
than silently ignoring them later.

Add tests for timestamp validation, and rename the test class to match
the name of the class under test.

Change-Id: I69f96c156840e018a70d5b8ad4e7ac66fada8eb8