SpecialAllEvents: preserve accordion state when form is submitted
Add a hidden form field that contains a comma-separated list of open
sections (any of 'upcoming' and 'ongoing'). Use the value of that field
to determine the initial state of the accordions.

On the client side, add a click event listener to the accordions and
update the hidden field as the accordions are collapsed/expanded.

Bug: T382248
Change-Id: I7c3aa08b5ee17001d38794fccdf6bb13a5a87df2