Add QUnit tests for EmailManager
Add plenty of test cases to cover all possible configurations for
participant selection, asserting that the displayed recipients are
correct, and also checking whether we warn about invalid recipients.

In some scenarios, we are not showing a warning when we should, or the
other way around. This is tracked in T366929. The respective test
assertions have been marked with a TODO. These TODOs also make a
distinction between "hard" and "soft" warning, to be interpreted in the
context of T366929 (with the soft warning being the one that reads
"may").

Change-Id: I9008a8333092541e04af4c72929ef3fc7dbb0280