Handle unattached global accounts in ListParticipantsHandler
If a global account is not attached on the current wiki, UserArray won't
create a User object for it, as the account doesn't really exist. This
code was already expecting null (see line 163), but mistakenly assumed
that the element would still be set in the array, while that is not the
case.

So, update the documentation and use null coalescing to avoid PHP
warnings.

Bug: T387914
Change-Id: I68de02169d8d413cbdc378de09adc35bd80db0eb