TimeZoneConverter: Avoid dependency on momentjs
Use the method described in T323193#10635272 to format the timestamp
using Intl. This is a bit convoluted, but hopefully we will have a core
library to handle it for us.

Drop the momentjs dependency as it's now unused.

Add a unit test to make sure that this works in common scenarios.

Change-Id: Ic014695b9992901262371b4f23a69a5e98b04326