Set proper merge strategy for CampaignEventsEventNamespaces
The default strategy for arrays is array_merge, but we want the value of
the configuration value to be overridden, not merged, consistently with
CommunityConfiguration. So, set the merge strategy to provide_default,
which effectively treats the extension.json value as a default that is
ignored in presence of an override.

Change-Id: Ibbb5a0ff55075be3d63354795bbb9edd957c31e2