Skip event page namespace validation on edits when page is unchanged
Make EventFactory accept a new validation flag that disables the
namespace check when the event page is unchanged. Callers are also
required to pass the previous event page if they want to opt-in for this
check. In theory this wouldn't be necessary; in practice, since
::newEvent takes a string, we would otherwise be duplicating the title
parsing logic in ::validatePage, which feels worse than adding a new
parameter.

Add a test with the various combinations of these parameters.

Use the new flag in Special:EditEventRegistration and
UpdateEventRegistration handler, so that the namespace check is skipped
as long as the user isn't changing the event page.

Bug: T387967
Bug: T387972
Change-Id: I22880307d5dd72e7befb7f40da0db516ef4f6f29