Add context parameter to UserLinker method
We want to make the UserLinker class use the newer UserLinkRenderer
instead of Linker. However, UserLinkRenderer::userLink takes a context
parameter, which we need to pass.

This patch makes our UserLinker methods take this context parameter and
updates callers. The parameter is still unused, but will become used in
a subsequent patch when Linker::userLinker is replaced.

Some callers are still falling back to the global state where proper
injection is difficult or not immediately needed.

Bug: T388427
Change-Id: I43d96f62cad1186c6bf09b7c01d947755abf04f3