TextWithIconWidget: return string instead of widget instance
Make the widget constructor private, and introduce a static `build`
method that returns a string. Update callers to use HtmlSnippet to keep
treating the string as raw HTML.

This is a first step towards converting the widget to Codex: unlike in
OOUI, Codex components are represented as strings (at least for the time
being). Callers should also not use OOUI for building simple HTML, which
will be done in a subsequent patch.

Bug: T388451
Change-Id: Iab7067c159a79923da80c2d938b488c1a38ba6a7