SpecialAllEvents: import TabLayout code from WikimediaCampaignEvents
Move TabLayout creation out of WikimediaCampaignEvents and let
SpecialAllEvents know about it, so that we can later allow users to
transclude individual tabs.

Introduce a new hook that passes an array of tabs that handlers can
append to, rather than a string of HTML that handlers are supposed
to wrap. Mark the new hook as internal, at least for the time being.
The old hook is now unused and will be dropped in a subsequent patch
(without deprecation).

See I00dd90802def176 for the corresponding removals in
WikimediaCampaignEvents.

Bug: T390083
Depends-On: I0706a22849a8581b87ae0f4c6a252bb7ebe95a03
Change-Id: If4b30423bee6d957958f6a99355e59397102a590