resources: Fix and enable no-jquery/no-done-fail
Use the recommended replacements of .then() and .catch(). In
ParticipantsManager, extract a long function used as done callback.

Update QUnit tests to wait for a `.then()` callback to run. This is
necessary due to the subtle difference between `.then()` and `.done()`
where the latter will run the callback synchronously if the promise is
already resolved when the callback is attached (as explained in
https://w.wiki/D$UN).

Change-Id: Idac8920aeabbec3ba03246a1f8a580b698949db0