Make CampaignEventsGetAllEventsTabsHook pass SpecialPage instance
The handler in WikimediaCampaignEvents needs to know whether the page is
being transcluded, so that it can change the output. So, change the
OutputPage parameter to pass the whole SpecialPage object, which
contains this information.

Note that passing a huge class such as SpecialPage around is not
desirable in general, but this specific hook is @internal and made for
this exact use case, so that should be fine.

Bug: T389892
Depends-On: I384c471e683ab6c112e71b4b6241826d33f67601
Change-Id: I67aff46d9321fc1580331806eb09ccca9f08f0df